requirements.txt
mouseinfo
Pillow
pygetwindow
pyautogui
keyboard